Add debug device.
Writing to Debug.stack prints a stack dump; writing to Debug.snapshot
makes a snapshot of the running VM and compares it to a previous one if
it exists; writing to Debug.exit terminates the VM; and writing to
Debug.test_mode activates features handy for automated testing.
test_mode is certainly subject to change.